# src/app/api/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dojo API",
    "description": "REST API for Dojo LMS — courses, chapters, purchases, promo codes, blog posts, and more. Uses OAuth2 with JWT access tokens and refresh token rotation.",
    "version": "0.1.0",
    "contact": {
      "name": "Dojo API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from POST /api/oauth/token"
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "password": {
            "tokenUrl": "/api/oauth/token",
            "scopes": {
              "read": "Read access to resources",
              "write": "Write access to resources",
              "admin": "Admin access"
            }
          },
          "clientCredentials": {
            "tokenUrl": "/api/oauth/token",
            "scopes": {
              "read": "Read access to resources",
              "write": "Write access to resources",
              "admin": "Admin access"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "Unauthorized" }
        },
        "required": ["error"]
      },
...
      "Success": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "clxyz123abc" },
          "email": {
            "type": "string",
            "format": "email",
            "example": "admin@dojo.api"
          },
          "name": { "type": "string", "example": "Admin User" },
          "role": {
            "type": "string",
            "enum": ["USER", "ADMIN"],
            "example": "ADMIN"
          },
          "isActive": { "type": "boolean", "example": true },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Course": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": {
            "type": "string",
            "example": "Introduction to TypeScript"
          },
          "slug": { "type": "string", "example": "introduction-to-typescript" },
          "description": { "type": "string", "nullable": true },
          "imageUrl": { "type": "string", "nullable": true },
          "price": { "type": "string", "example": "29.99", "nullable": true },
          "isPublished": { "type": "boolean" },
          "isListed": { "type": "boolean" },
          "isFeatured": { "type": "boolean" },
          "featuredOrder": { "type": "integer" },
          "outcomes": {
            "type": "string",
            "nullable": true,
            "description": "JSON array as text"
          },
          "requirements": {
            "type": "string",
            "nullable": true,
            "description": "JSON array as text"
          },
          "authorName": { "type": "string", "nullable": true },
          "authorRole": { "type": "string", "nullable": true },
          "categories": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Category" }
          },
          "chapters": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Chapter" }
          },
          "attachments": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Attachment" }
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CourseListItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "slug": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "imageUrl": { "type": "string", "nullable": true },
          "price": { "type": "string", "nullable": true },
          "isPublished": { "type": "boolean" },
          "categories": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Category" }
          },
          "_count": {
            "type": "object",
            "properties": {
              "chapters": { "type": "integer" },
              "purchases": { "type": "integer" }
            }
          },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "CoursePaginated": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CourseListItem" }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": { "type": "integer", "example": 1 },
              "limit": { "type": "integer", "example": 20 },
              "total": { "type": "integer", "example": 42 },
              "totalPages": { "type": "integer", "example": 3 }
            }
          }
        }
      },
      "Chapter": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string", "example": "Getting Started" },
          "description": { "type": "string", "nullable": true },
          "videoUrl": { "type": "string", "nullable": true },
          "timecodes": { "type": "string", "nullable": true },
          "notes": { "type": "string", "nullable": true },
          "homework": { "type": "string", "nullable": true },
          "position": { "type": "integer" },
          "isPublished": { "type": "boolean" },
          "isFree": { "type": "boolean" },
          "courseId": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Attachment": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string", "example": "syllabus.pdf" },
          "url": { "type": "string", "example": "/api/uploads/abc123.pdf" },
          "courseId": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Purchase": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "courseId": { "type": "string" },
          "amount": { "type": "string", "example": "0.00" },
          "promoCode": { "type": "string", "nullable": true },
          "course": {
            "type": "object",
            "properties": {
              "id": { "type": "string" },
              "title": { "type": "string" },
              "slug": { "type": "string" },
              "imageUrl": { "type": "string", "nullable": true }
            }
          },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string", "example": "Web Development" },
          "slug": { "type": "string", "example": "web-development" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/oauth/revoke": {
      "post": {
        "tags": ["OAuth2"],
        "operationId": "revokeToken",
        "summary": "Revoke a refresh token",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token"],
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "The refresh token to revoke"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token revoked",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Success" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Token not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/api/courses": {
      "get": {
        "tags": ["Courses"],
        "operationId": "listCourses",
        "summary": "List courses",
        "description": "Returns published courses for public users, all courses for admins. Supports pagination, search, and category filtering.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 20, "maximum": 50 }
          },
          {
            "name": "category",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by category slug"
          },
          {
            "name": "search",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Search in title and description"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated course list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CoursePaginated" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Courses"],
        "operationId": "createCourse",
        "summary": "Create a new course",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 3,
                    "example": "Advanced TypeScript"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Course created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Course" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden — admin only" }
        }
      }
    },
...
    "/api/courses/{courseId}/publish": {
      "patch": {
        "tags": ["Courses"],
        "operationId": "togglePublishCourse",
        "summary": "Toggle course publish status",
        "description": "Requires title, description, image, price, and at least one published chapter to publish.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Publish status toggled",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Course" }
              }
            }
          },
          "400": { "description": "Missing required fields or chapters" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Course not found" }
        }
      }
    },
    "/api/courses/{courseId}/chapters": {
      "post": {
        "tags": ["Chapters"],
        "operationId": "createChapter",
        "summary": "Add a chapter to a course",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string", "example": "Introduction" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Chapter created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Chapter" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" }
        }
      },
      "put": {
        "tags": ["Chapters"],
        "operationId": "reorderChapters",
        "summary": "Reorder chapters",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["list"],
                "properties": {
                  "list": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "position": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chapters reordered",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Success" }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/api/courses/{courseId}/attachments": {
      "post": {
        "tags": ["Attachments"],
        "operationId": "addAttachment",
        "summary": "Add an attachment to a course",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "url"],
                "properties": {
                  "name": { "type": "string", "example": "Cheat Sheet.pdf" },
                  "url": {
                    "type": "string",
                    "example": "/api/uploads/abc123.pdf"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Attachment created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Attachment" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/api/courses/{courseId}/purchase": {
      "post": {
        "tags": ["Purchases"],
        "operationId": "purchaseCourse",
        "summary": "Purchase a course",
        "description": "Mock payment — always succeeds. Optionally apply a promo code for a discount. Creates Purchase record, PromoCodeUsage, and increments counter in a single transaction.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "promoCode": { "type": "string", "example": "SUMMER50" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Purchase completed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Purchase" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Course not found or not published" },
          "409": { "description": "Already purchased" }
        }
      }
    },
...
    "/api/admin/users/{userId}": {
      "patch": {
        "tags": ["Admin"],
        "operationId": "updateUser",
        "summary": "Update a user",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "isActive": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "400": { "description": "No valid fields or update failed" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden — admin only" }
        }
      }
    },
...
  }
}